refactor: formattedCategories 함수 유틸로 분리 및 목록, 상세 페이지에 적용
