+++ frontend/src/components/common/ReviewCard/StrengthKeywordList/index.tsx
@@ -1,4 +1,5 @@
 import { Category } from '@/types';
+import formattedCategories from '@/utils/formattedCategories';
 
 import * as S from './styles';
 
@@ -7,22 +8,11 @@ interface StrengthKeywordProps {
 }
 
 const StrengthKeywordList = ({ categories }: StrengthKeywordProps) => {
-  const formattedCategories = categories.map((category) => {
-    // 카테고리 내용에서 '(예: ' 앞부분만 추출
-    const contentWithoutExample = Array.from(category.content.split(' (예: ')[0]);
-
-    const emoji = contentWithoutExample.shift();
-    const keyword = contentWithoutExample.join('');
-
-    return {
-      optionId: category.optionId,
-      content: `${emoji} ${keyword}`,
-    };
-  });
+  const transformedCategories = formattedCategories(categories);
 
   return (
     <S.KeywordList>
-      {formattedCategories.map((category) => (
+      {transformedCategories.map((category) => (
         <S.KeywordItem key={category.optionId}>{category.content}</S.KeywordItem>
       ))}
     </S.KeywordList>
